clean and modify test scripts
